+++ projects/angular-auth-oidc-client/src/lib/config/openid-configuration.ts
@@ -4,15 +4,15 @@ import { AuthWellKnownEndpoints } from './auth-well-known/auth-well-known-endpoi
 export interface OpenIdConfiguration {
   configId?: string;
   /**
-   * The url to the Security Token Service (STS) server.
+   * The url to the secure token server (STS) server. The authority issues tokens
    * This field is required.
    */
-  stsServer?: string;
-  /** Override the default STS wellknown endpoint postfix. */
+  authority?: string;
+  /** Override the default secure token server wellknown endpoint postfix. */
   authWellknownEndpointUrl?: string;
   authWellknownEndpoints?: AuthWellKnownEndpoints;
 
-  /** The redirect URL defined on the STS. */
+  /** The redirect URL defined on the secure token server. */
   redirectUrl?: string;
   /**
    * The Client MUST validate that the aud (audience) Claim contains its client_id value
@@ -29,9 +29,9 @@ export interface OpenIdConfiguration {
   responseType?: string;
   /**
    * This is this scopes which are requested from the server from this client.
-   * This must match the STS server configuration.
+   * This must match the secure token server configuration.
    * The 'openid' scope is required. The 'offline_access' scope can be requested when using refresh tokens
-   * but this is optional and some STS do not support this or recommend not requesting this even when using
+   * but this is optional and some secure token server do not support this or recommend not requesting this even when using
    * refresh tokens in the browser.
    */
   scope?: string;
@@ -73,7 +73,7 @@ export interface OpenIdConfiguration {
    * A token obtained by using a refresh token normally doesn't contain a nonce value.
    * However, some OIDC endpoint implementations do send one. The library checks to see if the nonce is present.
    * Note that if the nonce value is present, it will not be verified.
-   * This is not recommended, if the STS returns a nonce in the refresh.
+   * This is not recommended, if the secure token server returns a nonce in the refresh.
    * Default value is false
    */
   ignoreNonceAfterRefresh?: boolean;
@@ -98,7 +98,7 @@ export interface OpenIdConfiguration {
    * hard-configured routes (start, unauthorized, forbidden), this modification will add an extra
    * configuration option to override such behavior and trigger an event that will allow to subscribe to
    * it and let the application perform other actions. This would be useful to allow the application to
-   * save an initial return url so that the user is redirected to it after a successful login on the STS
+   * save an initial return url so that the user is redirected to it after a successful login on the secure token server
    * (ie: saving the return url previously on sessionStorage and then retrieving it during the triggering of the event).
    */
   triggerAuthorizationResultEvent?: boolean;