+++ projects/sample-code-flow-azuread/src/app/auth-config.module.ts
@@ -5,7 +5,7 @@ import { AuthModule, LogLevel } from 'angular-auth-oidc-client';
   imports: [
     AuthModule.forRoot({
       config: {
-        stsServer: 'https://login.microsoftonline.com/7ff95b15-dc21-4ba6-bc92-824856578fc1/v2.0',
+        authority: 'https://login.microsoftonline.com/7ff95b15-dc21-4ba6-bc92-824856578fc1/v2.0',
         authWellknownEndpointUrl: 'https://login.microsoftonline.com/common/v2.0',
         redirectUrl: window.location.origin,
         clientId: 'e38ea64a-2962-4cde-bfe7-dd2822fdab32',