+++ projects/sample-code-flow-http-config/src/app/auth-config.module.ts
@@ -10,7 +10,7 @@ export const httpLoaderFactory = (httpClient: HttpClient) => {
     .pipe(
       map((customConfig: any) => {
         return {
-          stsServer: customConfig.stsServer,
+          authority: customConfig.authority,
           redirectUrl: customConfig.redirect_url,
           clientId: customConfig.client_id,
           responseType: customConfig.response_type,